Track baseclasses in type propagation.
There is a new lattice point for each class, representing the values
that are subclasses of it.

To ensure that subclassing information is valued higher than subtype
information, we exploit that the ordering of lattice points determines
how the join operation chooses between ambiguous upper bounds.

BUG=
R=kustermann@google.com

Review URL: https://chromereviews.googleplex.com/488627013 .