Add convenience getter to extract the location of an AST node.
TreeNode.location is a getter for use in diagnostics and debugging.

It uses parent pointers to find the enclosing file URI and line ending
table of the node. If the node has no offset stored, it will use that
of the first enclosing parent that has an offset.

BUG=
R=ahe@google.com, jensj@google.com

Review URL: https://codereview.chromium.org/2448873004 .