Fixed crash when compiling without linking.
Since offsets and uris was added to the dill file, compiling to dill
without linking hasn't worked because of external libraries.

The external libraries doesn't add to the line starts table and thus
the uris from libraries, classes and so on isn't recorded.

When trying to write an unrecorded uri to the dill file it fails.

The simple solution applied here is to fall-back to the empty
string when the uri hasn't been recorded.

BUG=
R=asgerf@google.com

Review URL: https://codereview.chromium.org/2483243002 .